+++ imports/plugins/core/tax-base/client/index.js
@@ -1,6 +1,5 @@
 import "./checkout/taxes.html";
 import "./checkout/taxes.js";
-import "./settings/griddle.jsx";
 import "./settings/custom.html";
 import "./settings/custom.js";
 import "./settings/settings.html";
