+++ imports/plugins/core/ui-grid/client/index.js
@@ -1,0 +1,1 @@
+import "./griddle.jsx";
