+++ imports/plugins/core/ui-grid/register.js
@@ -1,0 +1,7 @@
+import { Reaction } from "/server/api";
+
+Reaction.registerPackage({
+  label: "UI Grid",
+  name: "reaction-ui-grid",
+  autoEnable: true
+});
