+++ private/data/i18n/en.json
@@ -150,19 +150,24 @@
         "paymentMethods": "Payment Methods",
         "externalServices": "External Services",
         "shopGeneralSettingsSaved": "Shop general settings saved.",
-        "shopGeneralSettingsFailed": "Shop general settings update failed. ",
+        "shopGeneralSettingsFailed": "Shop general settings update failed.",
         "shopAddressSettingsSaved": "Shop address settings saved.",
-        "shopAddressSettingsFailed": "Shop address settings update failed. ",
+        "shopAddressSettingsFailed": "Shop address settings update failed.",
         "shopMailSettingsSaved": "Shop mail settings saved.",
-        "shopMailSettingsFailed": "Shop mail settings update failed. ",
+        "shopMailSettingsFailed": "Shop mail settings update failed.",
         "shopExternalServicesSettingsSaved": "Shop external services settings saved.",
-        "shopExternalServicesSettingsFailed": "Shop external services settings update failed. ",
+        "shopExternalServicesSettingsFailed": "Shop external services settings update failed.",
         "shopLocalizationSettingsSaved": "Shop localization settings saved.",
-        "shopLocalizationSettingsFailed": "Shop localization settings update failed. ",
+        "shopLocalizationSettingsFailed": "Shop localization settings update failed.",
         "shopOptionsSettingsSaved": "Shop options saved.",
-        "shopOptionsSettingsFailed": "Shop options update failed. ",
+        "shopOptionsSettingsFailed": "Shop options update failed.",
         "shopPaymentMethodsSaved": "Shop Payment Methods settings saved.",
-        "shopPaymentMethodsFailed": "Shop Payment Methods settings update failed. "
+        "shopPaymentMethodsFailed": "Shop Payment Methods settings update failed."
+      },
+      "taxSettings": {
+        "noCustomTaxRatesFound": "No custom tax rates found.",
+        "shopCustomTaxRatesSaved": "Custom tax rate saved.",
+        "shopCustomTaxRatesFailed": "Error saving tax rate."
       },
       "header": {
         "tagsAdd": "Add tag",
